added final project

+++ PocketEngineer/app/src/main/java/com/example/pocketengineer/persistence/RoomLighting.kt
@@ -4,5 +4,5 @@ import java.io.Serializable
 
 data class RoomLighting(
         var roomName: String = "",
-        var lighting: Double = 0.0
+        var lighting: Int = 0
 ) : Serializable
